tools/mikrotik_preloader/requirements.txt
librouteros>=3.2.0
pyyaml>=6.0
